+++ modules/apps/site-initializer/site-initializer-raylife-ap/extra/remote-app/src/common/styles/components/activities.scss
@@ -98,9 +98,9 @@
 		font-weight: 700;
 
 		&.border-header {
-			border-top: 1px solid #ccdae8;
 			border-left: 0 !important;
 			border-right: 0 !important;
+			border-top: 1px solid #ccdae8;
 		}
 		&.border-bottom {
 			border-bottom: 1px solid #ccdae8;
